gym/requirements.txt
asgiref==3.8.1
asttokens==2.4.1
certifi==2024.8.30
charset-normalizer==3.3.2
crispy-bootstrap5==2024.2
decorator==5.1.1
Django==5.1.1
django-allauth==64.2.1
django-ckeditor==6.7.1
django-crispy-forms==2.3
django-flexible-subscriptions==0.15.1
django-grappelli==4.0.1
django-js-asset==2.2.0
django-taggit==6.0.0
django-tinymce==4.1.0
django-unfold==0.39.0
executing==2.1.0
idna==3.10
ipython==8.28.0
jedi==0.19.1
matplotlib-inline==0.1.7
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
